Reset Orientation, Fix debounce code
+ Added space bar trigger to reset oculus orientation
+ Working on the debounce code.  No switch is going to bounce
for 20 frames (that might be a full second at 20FPS.)

Ride is a lot smoother now, and is not as dependant on frame
rate.  HOWEVER there seems like there is an issue in the velocity
code, as it almost seems like pedaling faster makes you go slower or
the same.